requirements.txt
Django==5.0.3
djangorestframework==3.15.1
pillow==10.2.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
boto3==1.34.69
django-storages==1.14.2